Unbalanced One Factor

Unbalanced One Factor

Usage

balanced_one_factor(
  data,
  part = P,
  operator = O,
  measurement = Y,
  alpha = 0.05
)

Arguments

data

a data frame that contains measurements, operators and parts for a Gauge R&R analysis

part

a column of the data frame that has the part labels for the measurements

operator

a column of the data frame that has the operator labels for the measurements

measurement

a column of the data frame that has measurements of the object collected

alpha

the value for the confidence interval calculation (i.e. 95% CI would have alpha=0.05)

Value

a data frame of the values of the point estimates and the upper and lower bounds for each estimate

Examples

mydata <- data.frame(P=c(1,1,2,2,3,3),O=c(1,2,1,2,1,2),Y=c(5.3, 6.5, 5.4, 6.4, 6.9, 5.8))
balanced_one_factor(mydata, alpha=0.01)
#>   quantity   estimate       lower     upper
#> 1       mu  6.0500000   4.6909849  7.409015
#> 2     part -0.2062500 -10.9110737 10.883129
#> 3  gamma_m  0.5250000   0.1226812 21.959858
#> 4  gamma_r -0.3928571  -0.4978485 20.839261